+++ src/device/tsp.script/Script/ScriptEntityBaseCollection.cs
@@ -2,7 +2,6 @@ using System.Collections.ObjectModel;
 using System.Diagnostics;
 using cc.isr.Std.TrimExtensions;
 using cc.isr.VI.Pith;
-using cc.isr.VI.Syntax.Tsp;
 using cc.isr.VI.Tsp.Script.SessionBaseExtensions;
 
 namespace cc.isr.VI.Tsp.Script;