+++ src/device/tsp.script/Script/SessionBaseExtensions/FirmwareManager-delete.cs
@@ -1,4 +1,3 @@
-using System;
 using cc.isr.VI.Pith;
 
 namespace cc.isr.VI.Tsp.Script.SessionBaseExtensions;